use router to make the app more cleaner
